src-tauri/resources/near-intents-helper/requirements.txt
requests>=2.31,<3
PyNaCl>=1.5,<2
python-dotenv>=1.0,<2
base58>=2.1,<3
borsh-construct>=0.1.0,<1
